add data warehouse source tables into pipeline ui
